k2/requirements.txt
addons/k2-addon
addons/k2-connectioncheck
addons/k2-consolerunner
addons/k2-gitrepo
addons/k2-gtestbinaryrunner
addons/k2-healthcheck
addons/k2-infra
addons/k2-jenkins
addons/k2-linuxutils
addons/k2-looper
addons/k2-mender
addons/k2-metrics
addons/k2-monitor
addons/k2-multirunner
addons/k2-networking
addons/k2-orchestration
addons/k2-power
addons/k2-profiling
addons/k2-serial
addons/k2-sispmpowerswitch
addons/k2-sutevents
addons/k2-telnet
addons/k2-upgrade
addons/k2-zelenium
numpy==1.19.4 --only-binary numpy
arrow==0.12.1
docker-pycreds==0.4.0
docker==3.7.0
fastentrypoints==0.10
GitPython==2.1.9
httpretty==1.0.3
jinja2==2.9.6
junit-xml==1.8
matplotlib==3.3.3 --only-binary matplotlib
nose-timer==0.7.3
nose==1.3.7
pandas==1.1.4 --only-binary pandas
pretenders==1.4.4
pyserial==3.5
python-jenkins==1.4.0
pytracing==0.2
selenium==3.9.0
simple-websocket-server==0.4.0
testinfra==1.16.0 --only-binary testinfra
thrift==0.10.0
wakeonlan==1.1.6
websocket-client==0.55.0
zenterio-zaf
